Development

+++ lib/layouts/home/tabs/home_tab/widgets/advertisements/provider/adv_provider.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class AdvProvider extends ChangeNotifier {
-  double currentIndex = 0;
-  changeAdvCurrentIndex(double newValue) {
+  int currentIndex = 0;
+  changeAdvCurrentIndex(int newValue) {
     if (currentIndex == newValue) return;
     currentIndex = newValue;
     notifyListeners();
